Add scrolling text in multiple views
Now we can add multiple documents in at a time and scroll the text
around by clicking on the slider.  Scrolling one document does not
effect other documents, making rendering faster.

Also abandoned the linear search strategy for finding lines, and added
in a binary search for a suitable line.  It remains to be seen if this
will give us reasonable scrolling performance.