Better scroll performance for vector fonts
Previously, we rendered each word invidiually, which gave us a lot of
calls to RFont.draw() and translate().  Now I build up each line in
memory, and render the line when we think we're going off screen.  This
is as far as we can go because the RFont.draw() call doesn't understand
endlines or carriage returns.

The text still looks terrible, but this might be part of the particular
font or settings I'm using.