First pass at making hierarchy a relative concept
In order to support two separate hierarchies with the same nodes, I've
changed a lot of concepts that used to be defined in terms of absolute
hierarchy level to being defined in terms of the relative height of node
in the tree.  Height has the advantage that it does not stipulate what
the maximum hypothetical height is.

WordNodes do not support dual membership yet, they still believe
themselves to be members of exactly one tree.  When I add the XML parser
to construct the new tree, it will become more apparent what sort of
interface we'll need to support dual word membership.